[MemoryAllocator] Align buffers to size_t size

+++ lib/CodeGen/MemoryAllocator.cpp
@@ -10,7 +10,7 @@ const size_t MemoryAllocator::npos = -1;
 
 size_t MemoryAllocator::allocate(size_t size) {
   // Always allocate buffers properly aligned to hold values of any type.
-  size = (size + alignof(float) - 1) & ~(alignof(float) - 1);
+  size = (size + alignof(size_t) - 1) & ~(alignof(size_t) - 1);
   size_t prev = 0;
   for (auto it = allocations_.begin(), e = allocations_.end(); it != e; it++) {
     if (it->begin_ - prev >= size) {
